Finish dev-net refactoring

+++ deployments/dev-net/dn-build.proxy.sh
@@ -1,30 +1,34 @@
-#!/usr/bin/env bash
+#!/bin/bash
 
-#################################
+#######################################
 # Build proxy docker image
-#################################
+#######################################
+
+if [[ ( $# -eq 1 ) && ( $1 = "-h" || $1 = "--help" ) ]]; then
+  echo "Usage: ./dev-net-build.proxy.sh"
+  exit 0
+fi
 
 . ./scripts/setup.sh
 
 SECONDS=0
 
+./cleanup.sh
+
 printf "Cleaning up old image $PROXY_IMAGE\n"
 docker rmi $PROXY_IMAGE
 
 printf "Creating build context of proxy server ...\n"
-cd $ROOT_DIR/packages/proxy
-yarn tsc:prod
-printMessage "Create tsc for proxy server"  $?
+./build-proxy.sh
 
-yarn pack:lib
 printMessage "Create build context for proxy server"  $?
 sleep 1
 
 ### build image ###
+cd $ROOT_DIR/packages/proxy
 DOCKER_BUILD=1 docker build --no-cache -t $PROXY_IMAGE .
 printMessage "Create proxy image" $?
 sleep 1
 
 duration=$SECONDS
 printf "${GREEN}$(($duration / 60)) minutes and $(($duration % 60)) seconds elapsed.\n\n${NC}"
-
